+++ pages/candidates/components/Blog/BlogCard/index.tsx
