+++ pages/candidates/components/Blog/BlogCard/styles.ts
