+++ pages/candidates/components/Blog/BlogSection/index.tsx
