+++ pages/candidates/components/Blog/BlogSection/styles.ts
