+++ pages/candidates/components/Blog/BlogSwiper/index.tsx
