+++ pages/candidates/components/Blog/BlogSwiper/styles.ts
