+++ pages/candidates/components/CallToAction/styles.ts
