+++ pages/candidates/components/Community/SocialCard/index.tsx
