+++ pages/candidates/components/Community/SocialCard/styles.ts
