+++ pages/candidates/components/Community/index.tsx
