+++ pages/candidates/components/Community/styles.ts
