+++ pages/candidates/components/DiscoverGA/index.tsx
