+++ pages/candidates/components/DiscoverGA/style.ts
