+++ pages/candidates/components/OurServices/OurServicesCard/styles.ts
