+++ pages/candidates/components/OurServices/OurServicesSwiper/styles.ts
