+++ pages/candidates/components/Success-stories/style.ts
