+++ pages/candidates/components/WorkIn/WorkInCard.tsx
