+++ pages/candidates/components/WorkIn/styles.ts
