+++ pages/components/HeroSection/index.tsx
