+++ pages/components/HeroSection/styles.ts
