+++ pages/components/MeetTheFoundersSection/index.tsx
