+++ pages/components/MeetTheFoundersSection/styles.ts
