+++ pages/components/PartnersSection/index.tsx
