+++ pages/components/PartnersSection/styles.ts
