+++ pages/components/SectionDivider/index.tsx
