+++ pages/components/SectionDivider/styles.ts
