+++ pages/index.tsx
@@ -1,48 +1,28 @@
-import CallToAction from './components/CallToAction/CallToAction';
-import WorkIn from '@pages/components/WorkIn/WorkIn';
-import OurServicesSection from '@pages/components/OurServices';
-import SuccessStories from '@pages/components/Success-stories';
-import CommunitySection from '@pages/components/Community';
-import BlogSection from '@pages/components/Blog/BlogSection';
+import React from 'react';
+
+import { Container, Title } from './styles';
+import HeroSection from './components/HeroSection';
 import { ReactElement } from 'react';
+import { PageLayout } from '@/components';
+import ConsultingGmbHSection from './components/ConsultingGmbHSection/index';
+import RecruitingAndCareerServicesSection from './components/RecrutingAndCareerServicesSection';
+import PartnersSection from './components/PartnersSection';
+import SectionDivider from './components/SectionDivider';
+import { MeetTheFoundersSection } from './components/MeetTheFoundersSection';
 
-import { NextPageWithLayout } from '@pages/_app';
-import PageLayout from '@components/PageLayout';
-import DiscoverGA from '@pages/components/DiscoverGA';
-import { useRouter } from 'next/router';
-import { IntlProvider } from 'react-intl';
-import de from '../lang/de.json';
-import en from '../lang/en.json';
-import pt from '../lang/pt.json';
-
-const Page: NextPageWithLayout = () => {
+export default function Main() {
   return (
     <>
-      <CallToAction />
-      <WorkIn />
-      <DiscoverGA />
-      <OurServicesSection />
-      <SuccessStories />
-      <CommunitySection />
-      <BlogSection />
+      <HeroSection />
+      <ConsultingGmbHSection />
+      <RecruitingAndCareerServicesSection />
+      <PartnersSection />
+      <SectionDivider />
+      <MeetTheFoundersSection />
     </>
   );
+}
+
+Main.getLayout = function getLayout(page: ReactElement) {
+  return <PageLayout>{page}</PageLayout>;
 };
-
-export const languages = {
-  de,
-  en,
-  pt,
-};
-
-Page.getLayout = function getLayout(page: ReactElement) {
-  const { locale } = useRouter();
-
-  return (
-    <IntlProvider locale={locale!} messages={languages[locale]}>
-      <PageLayout>{page}</PageLayout>
-    </IntlProvider>
-  );
-};
-
-export default Page;
