+++ pages/styles.ts
